Feature/saas 338 Refactor list search template to directive

See merge request !295

+++ app/scripts/controllers/backups-controller.js
@@ -6,15 +6,14 @@
       'backupsService',
       'baseControllerListClass',
       '$state',
+      'ENTITYLISTFIELDTYPES',
       BaseBackupListController
     ]);
 
-  function BaseBackupListController(backupsService, baseControllerListClass, $state) {
-    var controllerScope = this;
+  function BaseBackupListController(backupsService, baseControllerListClass, $state, ENTITYLISTFIELDTYPES) {
     var Controller = baseControllerListClass.extend({
       init:function() {
         this.service = backupsService;
-        this.controllerScope = controllerScope;
         this._super();
         this.searchFieldName = 'description';
         this.actionButtonsListItems = [
@@ -27,11 +26,48 @@
             clickFunction: this.deleteBackup.bind(this.controllerScope)
           }
         ];
+        this.entityOptions = {
+          entityData: {
+            title: 'Backups',
+            createLink: 'backups.create',
+            createLinkText: 'Create a backup',
+            noDataText: 'You have no backups yet.',
+            expandable: false
+
+          },
+          list: [
+            {
+              name: 'Description',
+              propertyName: 'description',
+              type: ENTITYLISTFIELDTYPES.link,
+              link: '',
+              showForMobile: ENTITYLISTFIELDTYPES.showForMobile
+            },
+            {
+              name: 'State',
+              propertyName: 'state',
+              type: ENTITYLISTFIELDTYPES.noType
+            },
+            {
+              name: 'Kept',
+              propertyName: 'kept_until',
+              emptyText: 'kept forever',
+              type: ENTITYLISTFIELDTYPES.noType
+            },
+            {
+              name: 'Date',
+              propertyName: 'created_at',
+              type: ENTITYLISTFIELDTYPES.date
+            }
+          ]
+        };
       },
       deleteBackup:function(backup) {
         var vm = this;
-        vm.service.deleteBackup(backup.uuid).then(
-          vm.getList, vm.handleActionException);
+        if (confirm("Confirm deletion")) {
+          vm.service.deleteBackup(backup.uuid).then(
+            vm.getList, vm.handleActionException);
+        }
       },
       restoreBackup:function(backup) {
         $state.go('backups.restore', {uuid: backup.uuid});
@@ -53,6 +89,10 @@ angular.module('ncsaas')
   function BackupListController(BaseBackupListController) {
     var controllerScope = this;
       var Controller = BaseBackupListController.extend({
+        init:function() {
+          this.controllerScope = controllerScope;
+          this._super();
+        }
       });
 
     controllerScope.__proto__ = new Controller();
